# doc/swaggers/swagger_v1.1.yaml
openapi: 3.0.3
info:
  title: Repo API
  description: API for managing posts, replies, and social interactions in a MongoDB-backed system.
  version: 1.0.0
servers:
  - url: https://api.example.com
paths:
  /post:
    post:
      summary: Create a new post
      operationId: createPost
      tags:
      - Post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /post/{id}:
    get:
      summary: Get a specific post
      operationId: getPost
      tags:
      - Post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    put:
      summary: Edit a post
      operationId: editPost
      tags:
      - Post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditPostRequest"
      responses:
        "200":
          description: Post updated successfully
    delete:
      summary: Delete a post
      operationId: deletePost
      tags:
      - Post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Post deleted successfully
  /post/{id}/reply:
    post:
      summary: Reply to a post
      operationId: replyToPost
      tags:
        - Post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Reply created successfully
    get:
      summary: Get all replies for a post
      operationId: getPostReplies
      tags:
      - Post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of replies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
          format: objectId
        ownerId:
          type: string
          format: objectId
        text:
          type: string
        media:
          type: string
          description: Link to S3
        repostId:
          type: string
          format: objectId
          nullable: true
        replyToPostId:
          type: string
          format: objectId
          nullable: true
        isReply:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        deleted:
          type: boolean
          default: false
...
    CreatePostRequest:
      type: object
      required:
        - ownerId
        - text
      properties:
        ownerId:
          type: string
          format: objectId
        text:
          type: string
        media:
          type: string
          nullable: true
    EditPostRequest:
      type: object
      properties:
        text:
          type: string
        media:
          type: string
    CreateReplyRequest:
      type: object
      required:
        - ownerId
        - text
      properties:
        ownerId:
          type: string
          format: objectId
        text:
          type: string